requirements.txt
streamlit>=1.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pathlib>=1.0.0
requests>=2.31.0